requirements.txt
discord
requests
aiohttp
dataclasses
PyYAML